Implemented second algorithm.

+++ src/algorithmsPackage/Algorithm.java
@@ -2,5 +2,5 @@ package algorithmsPackage;
 
 public interface Algorithm {
 
-    float getAverageWaitingTime();
+    double getAverageWaitingTime();
 }
